Remove check when extracting smiles
